Add files via upload
index.html = added <h2> which specifies in which page are we.
firstProgram.html =  added <h2> which specifies in which page are we.
variables.html = added <h2> which specifies in which page we are. added a "nothing to present yet" paragraph. 
printAndScan.html = added <h2> which specifies in which page we are. added content before editing code.
mathOperator.html = added <h2> which specifies in which page we are. added content before editing code.
statments.html = added <h2> which specifies in which page we are. added content before editing code.
styles.css = added a few versions of the background page (bg, bg-oneAndHalf, bg-two) - in order to determine if we need a page, a page and a half , or two pages.